updated plots to make the axes equal.
Still have a bug where the axes on the thetas are getting set
to the same scale of those in mm. This collapses all the points to just a
single dot around zero. I know what the problem is, just not sure how I want
to fix it.

Now I'm adding what moves would look like starting first with constant
linear acceleration moves.